New module for functions to check prisoner location
This is to break the dependency from some `utils/utils`
to `data/offenderSearchApi` which ends up importing
`bynyan` which doesn't work anymore with Cypress 15.x
(because it uses Node's `util` in a brower environment)
