+++ server/utils/utils.ts
@@ -1,4 +1,5 @@
-import { isBeingTransferred, isOutside, type OffenderSearchResult } from '../data/offenderSearchApi'
+import { type OffenderSearchResult } from '../data/offenderSearchApi'
+import { isOutside, isBeingTransferred } from './offenderSearchUtils'
 
 const properCase = (word: string): string =>
   word.length >= 1 ? word[0].toUpperCase() + word.toLowerCase().slice(1) : word
